requirements.txt
alembic==1.8.1
django_recaptcha==3.0.0
Flask==2.2.3
Flask_Mail==0.9.1
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
numpy==1.21.5
opencv_python==4.5.5.64
Pillow==9.4.0
SQLAlchemy==1.4.39
torch==1.9.1+cu102
torchvision==0.10.1+cu102
Werkzeug==2.2.3
WTForms==3.0.1
